pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "socket-firewall-configurator"
version = "1.0.0"
description = "Automated socket firewall configuration generator"
readme = "README.md"
license = {text = "BSD-3-Clause"}
requires-python = ">=3.10"
authors = [
    {name = "Infrastructure Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: System :: Networking :: Firewalls",
]
dependencies = [
    "PyYAML>=6.0.1",
    "rich>=13.7.0",
    "requests>=2.31.0",
    "jsonschema>=4.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "flake8>=7.0.0",
    "black>=24.1.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "bandit>=1.7.0",
    "safety>=2.3.0",
    "pre-commit>=3.6.0",
    "types-PyYAML>=6.0.0",
    "types-requests>=2.31.0",
]

[project.scripts]
socket-firewall-configurator = "socket_firewall_configurator.configurator:main"

[tool.setuptools.packages.find]
where = ["app"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".git", ".venv", "venv", "build", "dist"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = ["tests", "venv", ".venv"]

[[tool.mypy.overrides]]
module = "yaml.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101"]

